refactor(package): updated depensort to use yargs package
Updated the package to use the yargs package to enable us to easily update/add features

BREAKING CHANGE: updated the entry point to use yargs

fixes #19
