+++ .prettierrc
@@ -10,6 +10,6 @@
 	"arrowParens": "avoid",
 	"requirePragma": false,
 	"singleAttributePerLine": true,
-	"endOfLine": "lf",
+	"endOfLine": "crlf",
 	"useTabs": true
 }
